requirements.txt
kivy
kivymd
psycopg2
bcrypt
python-dotenv
requests
Pillow
spotipy
